Move get_device_info into the device module.
