+++ kolibri/core/public/utils.py
@@ -2,7 +2,6 @@ import datetime
 import hashlib
 import json
 import logging
-import platform
 import random
 
 import requests
@@ -16,13 +15,11 @@ from requests.exceptions import ConnectionError
 from rest_framework import status
 from six.moves.urllib.parse import urljoin
 
-import kolibri
 from kolibri.core.auth.constants.morango_sync import PROFILE_FACILITY_DATA
 from kolibri.core.auth.constants.morango_sync import ScopeDefinitions
 from kolibri.core.auth.models import FacilityUser
 from kolibri.core.device.models import UserSyncStatus
-from kolibri.core.device.utils import DeviceNotProvisioned
-from kolibri.core.device.utils import get_device_setting
+from kolibri.core.device.utils import get_device_info
 from kolibri.core.public.constants.user_sync_statuses import QUEUED
 from kolibri.core.public.constants.user_sync_statuses import SYNC
 from kolibri.core.tasks.api import prepare_soud_resume_sync_job
@@ -37,65 +34,6 @@ from kolibri.utils.conf import OPTIONS
 
 
 logger = logging.getLogger(__name__)
-
-
-device_info_keys = {
-    "1": [
-        "application",
-        "kolibri_version",
-        "instance_id",
-        "device_name",
-        "operating_system",
-    ],
-    "2": [
-        "application",
-        "kolibri_version",
-        "instance_id",
-        "device_name",
-        "operating_system",
-        "subset_of_users_device",
-    ],
-}
-
-DEVICE_INFO_VERSION = "2"
-
-
-def get_device_info(version=DEVICE_INFO_VERSION):
-    """
-    Returns metadata information about the device
-    The default kwarg version should always be the latest
-    version of device info that this function supports.
-    We maintain historic versions for backwards compatibility
-    """
-
-    if version not in device_info_keys:
-        version = DEVICE_INFO_VERSION
-
-    instance_model = InstanceIDModel.get_or_create_current_instance()[0]
-    try:
-        device_name = get_device_setting("name")
-        subset_of_users_device = get_device_setting("subset_of_users_device")
-    # When Koliri starts at the first time, and device hasn't been created
-    except DeviceNotProvisioned:
-        device_name = instance_model.hostname
-        subset_of_users_device = False
-
-    all_info = {
-        "application": "kolibri",
-        "kolibri_version": kolibri.__version__,
-        "instance_id": instance_model.id,
-        "device_name": device_name,
-        "operating_system": platform.system(),
-        "subset_of_users_device": subset_of_users_device,
-    }
-
-    info = {}
-
-    # By this point, we have validated that the version is in device_info_keys
-    for key in device_info_keys.get(version, []):
-        info[key] = all_info[key]
-
-    return info
 
 
 def find_soud_sync_sessions(using=None, **filters):
